pyproject.toml
[tool.poetry]
name = "fastapi-board"
version = "0.1.0"
description = "Board Take-Home Assignment"
authors = ["Geonhyuk Im <athanasia9410@gmail.com>"]
readme = "README.md"
packages = [{include = "fastapi_board"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "0.104.1"
uvicorn = "0.24.0"
SQLAlchemy = "2.0.23"
passlib = "^1.7.4"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.9"
pydantic = {extras = ["email"], version = "^2.5.2"}
pyjwt = "^2.8.0"
python-multipart = "^0.0.6"
redis = "^5.0.1"
fastapi-middleware = "^0.1.0"
itsdangerous = "^2.1.2"
pytz = "^2023.3.post1"

[tool.poetry.group.dev.dependencies]
pytest = "7.4.3"
alembic = "1.13.0"
mypy = "1.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
